+++ package.json
@@ -15,8 +15,11 @@
     "@next/mdx": "^13.0.7",
     "date-fns": "^2.29.3",
     "disqus-react": "^1.1.5",
+    "dom-serializer": "^2.0.0",
+    "domutils": "^3.0.1",
     "gitlog": "^4.0.4",
     "gray-matter": "^4.0.3",
+    "htmlparser2": "^8.0.1",
     "interweave": "^13.0.0",
     "next": "^13.0.3",
     "next-mdx-remote": "^4.2.0",
